Fix: Complie Warning Message

+++ ioManager.cpp
@@ -65,6 +65,6 @@ void IOManager::clear()
 
 void IOManager::gotoxy(int x, int y)
 {
-	COORD pos = { x, y };
+	COORD pos = { (short)x, (short)y };
 	SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), pos);
 }
